+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Tideland Go Library
 
+## 2017-01-07
+
+- *etc.Etc* now can write to an *io.Reader*
+
 ## 2016-12-11
 
-- Added asserts for *range*, *cae*, and *path exists*
+- Added asserts for *range*, *case*, and *path exists*
 
 ## 2016-12-07